requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.3
data==0.4
decorator==4.4.2
funcsigs==1.0.2
iniconfig==1.0.1
linecache2==1.0.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.0
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.1
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.25.10
